% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erpcore-package.R
\docType{package}
\name{erpcore-package}
\alias{erpcore}
\alias{erpcore-package}
\title{erpcore: Download ERP CORE Data}
\description{
Downloads data from the ERP Core datasets, currently stored on the 
    Open Science Framework (https://www.osf.io)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eegverse/erpcore}
  \item Report bugs at \url{https://github.com/eegverse/erpcore/issues}
}

}
\author{
\strong{Maintainer}: Matt Craddock \email{matt@mattcraddock.com} [copyright holder]

Other contributors:
\itemize{
  \item Emily S. Kappenman [copyright holder]
  \item Jaclyn L. Farrens [copyright holder]
  \item Wendy Zhang [copyright holder]
  \item Andrew X. Stewart [copyright holder]
  \item Steven J. Luck [copyright holder]
}

}
\keyword{internal}
