% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_erpcore.R
\name{get_erpcore}
\alias{get_erpcore}
\alias{get_n170}
\alias{get_mmn}
\alias{get_n2pc}
\alias{get_n400}
\title{Download ERP Core Components}
\usage{
get_erpcore(
  component = c("n170", "mmn", "n2pc", "n400"),
  dest_path = NULL,
  conflicts = "error",
  type = c("raw", "bids", "all")
)

get_n170(dest_path = NULL, conflicts = "error", type = c("raw", "bids", "all"))

get_mmn(dest_path = NULL, conflicts = "error", type = c("raw", "bids", "all"))

get_n2pc(dest_path = NULL, conflicts = "error", type = c("raw", "bids", "all"))

get_n400(dest_path = NULL, conflicts = "error", type = c("raw", "bids", "all"))
}
\arguments{
\item{component}{Options are:
\itemize{
\item "n170": download data for the face-specific N170 component.
\item "mmn": download data for the Mismatch Negativity component.
\item "n2pc": download data for the N2pc spatial attention component.
\item "n400": download data for the N400 semantic mismatch component.
}}

\item{dest_path}{destination file path}

\item{conflicts}{This determines what happens when a file with the same name exists at the specified destination. Can be one of the following:
\itemize{
\item "error" (the default): throw an error and abort the file transfer operation.
\item "skip": skip the conflicting file(s) and continue transferring the remaining files.
\item "overwrite": replace the existing file with the transferred copy.
}}

\item{type}{The data is in three different formats:
\itemize{
\item "raw": download the raw data and scripts.
\item "bids": download the raw data in scripts in a BIDS compatible format.
\item "all": download the raw and processed data and scripts.
}}
}
\description{
Download ERP Core Components
}
\section{Functions}{
\itemize{
\item \code{get_n170}: Retrieve N170 data

\item \code{get_mmn}: Retrieve MMN data

\item \code{get_n2pc}: Retrieve N2pc data

\item \code{get_n400}: Retrieve N400 data
}}

