% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_erpcore.R
\name{get_mmn}
\alias{get_mmn}
\title{Download ERP Core MMN}
\usage{
get_mmn(dest_path = NULL, file_name = NULL, overwrite = FALSE)
}
\arguments{
\item{dest_path}{destination file path}

\item{file_name}{destination file name}

\item{overwrite}{overwrite existing files (defaults to FALSE)}
}
\description{
Download ERP Core MMN
}
