% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_erpcore.R
\name{get_N170}
\alias{get_N170}
\title{Download ERP Core N170}
\usage{
get_N170(dest_path = NULL, file_name = NULL)
}
\arguments{
\item{dest_path}{destination file path}

\item{file_name}{destination file name}
}
\description{
Download ERP Core N170
}
