% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_erpcore.R
\name{get_n2pc}
\alias{get_n2pc}
\title{Download ERP Core n2pc}
\usage{
get_n2pc(dest_path = NULL, file_name = NULL, overwrite = FALSE)
}
\arguments{
\item{dest_path}{destination file path}

\item{file_name}{destination file name}

\item{overwrite}{overwrite existing files (defaults to FALSE)}
}
\description{
Download ERP Core n2pc
}
