% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_erpcore.R
\name{get_n400}
\alias{get_n400}
\title{Download ERP Core N400}
\usage{
get_n400(dest_path = NULL, file_name = NULL, overwrite = FALSE)
}
\arguments{
\item{dest_path}{destination file path}

\item{file_name}{destination file name}

\item{overwrite}{overwrite existing files (defaults to FALSE)}
}
\description{
Download ERP Core N400
}
